requirements.txt
numpy<2
torch
torchvision
ultralytics
opencv-python
pyttsx3
pygame
sort-tracker
